added web-tab iconography and some MAP settings
The Icon itself is the Japanese name of Tachibana Doetsu's Legendary blade, Chidori (Thousand Birds). Albeit, it does not look so well(it was done by hand with a mouse). You can barely make out the word bird if you look at the actual image. Its also strange that it looks like "TC" when in the browser, I dont think that will ever improve. But hey, it is better than seeing a simple stock file icon right?